test(formal-context): exercise rejection through dispatch
