+++ tests/math/formal_concept_analysis/test_formal_concept_analysis.py
@@ -5,8 +5,6 @@ from __future__ import annotations
 import pytest
 from pydantic import ValidationError
 
-from jacobian.catalog.catalog import Catalog
-from jacobian.dispatch import OperationRequestValidationError, invoke_operation
 from jacobian.math.formal_concept_analysis import FormalContext
 from jacobian.math.formal_concept_analysis._models import (
     AttributeSubsetRequest,
@@ -187,45 +185,6 @@ class TestConceptLattice:
 
 
 class TestValidation:
-    @pytest.mark.parametrize(
-        ("operation_id", "context"),
-        (
-            (
-                "formal_context.attributes.derivation.compute",
-                {
-                    "objects": ["o0", "o1"],
-                    "attributes": ["a0"],
-                    "incidence": [[0, 0], [1, 0]],
-                },
-            ),
-            (
-                "formal_context.concept.from_attributes.compute",
-                {
-                    "objects": ["o0", "o1"],
-                    "attributes": ["a0"],
-                    "incidence": [[0, 0], [1, 0]],
-                },
-            ),
-            (
-                "formal_context.objects.derivation.compute",
-                {
-                    "objects": ["o0"],
-                    "attributes": ["a0", "a1"],
-                    "incidence": [[0, 0], [0, 1]],
-                },
-            ),
-        ),
-    )
-    def test_public_operations_reject_indices_outside_their_axis(
-        self, operation_id: str, context: dict[str, object]
-    ) -> None:
-        with pytest.raises(OperationRequestValidationError):
-            invoke_operation(
-                operation_id,
-                {"context": context, "subset": [1]},
-                Catalog.open(),
-            )
-
     def test_duplicate_objects_rejected(self) -> None:
         with pytest.raises(ValidationError, match="unique"):
             FormalContext(
